pyproject.toml
[tool.poetry]
name = "whisper-input-mac"
version = "0.1.0"
description = "A macOS menu bar app for Whisper input"
authors = ["Abhigya Maskay <ave70011@gmail.com>"]
readme = "README.md"
packages = [{include = "whisper_input_mac", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pyobjc-core = "^11.0"
pyobjc-framework-cocoa = "^11.0"
pyobjc-framework-ApplicationServices = "^11.0"
pyobjc-framework-Quartz = "^11.0"
pyobjc-framework-AVFoundation = "^11.0"
lightning-whisper-mlx = "^0.0.10"
ml-dtypes = "^0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21"

[tool.poetry.scripts]
whisper-input-mac = "whisper_input_mac.app:main"
whisper-prefetch-models = "whisper_input_mac.tools.prefetch_models:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
